remove console log & add guard

+++ src/services/apiCabins.js
@@ -1,7 +1,6 @@
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function createEditCabin(newCabin, id) {
-  console.log(newCabin, id);
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
@@ -29,6 +28,8 @@ export async function createEditCabin(newCabin, id) {
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
+
+  if (hasImagePath) return data;
 
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data[0].id);
